refactor(service): update imports to work with new project structure
