+++ service/tests/conftest.py
@@ -2,9 +2,9 @@ import numpy as np
 import pytest
 from fastapi.testclient import TestClient
 
-from ..main import app
-from ..models.model import Equations, RMRModel
-from ..models.time_projection import TimeProjection
+from main import app
+from models.model import Equations, RMRModel
+from models.time_projection import TimeProjection
 
 
 @pytest.fixture
